+++ src/ocpp/Messages.cpp
@@ -36,7 +36,7 @@ size_t ICall::measureJson() const {
     return this->serializeJson(nullptr, 0);
 }
 
-static uint32_t next_call_id = 0;
+static uint64_t next_call_id = 0;
 
 const char * const ChangeAvailabilityResponseStatusStrings[] = {
     "Accepted",
